HPCC-25484 WIP: Create daliadmin helper functions for ESP WsDali
1. Create 2 new files in dali/daliadmin: daadmin.cpp and daadmin.hpp.
2. Move the code from daliadmin.cpp to daadmin.cpp except for main(),
usage(), and a few lines needed by main().
3. For the helper functions used inside the main(), add the 'extern
da_decl' declarations into the daadmin.hpp.
4. For those helper functions in daadmin.cpp, remove the 'static' and
the default input values if any.
